add license and move to plugin examples folder
