+++ Examples/other_plugins/InjectTraits.swift
@@ -1,0 +1,54 @@
+//
+//  InjectTraits.swift
+
+//  Created by Alan Charles on 12/7/22.
+//
+
+// This plugin is NOT SUPPORTED by Segment.  It is here merely as an example,
+// and for your convenience should you find it useful.
+
+// MIT License
+//
+// Copyright (c) 2021 Segment
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import Foundation
+import Segment
+
+class InjectTraits: Plugin {
+    let type = PluginType.enrichment
+    weak var analytics: Analytics? = nil
+    
+    func execute<T: RawEvent>(event: T?) -> T? {
+        if event?.type == "identify" {
+            return event
+        }
+        
+        var workingEvent = event
+        
+        if var context = event?.context?.dictionaryValue {
+            context[keyPath: "traits"] = analytics?.traits()
+            
+            workingEvent?.context = try? JSON(context)
+        }
+        
+        return workingEvent
+    }
+}
